AI Chatbot In Python
This automated communication system is developed using Python. The project file contains a python script (main.py, trainingData.py, JSON file, and pkl file). Talking about this chatbot, it allows the user to provide suitable queries about the college and replies with suitable answers. You change the data by changing the given JSON file. Also, this is a simple cmd-based project which is easy to understand and use.